harness: absorb closure audit into mcp
